+++ Mods ChoGGi/All Sponsor Buildings/metadata.lua
@@ -1,5 +1,6 @@
 return PlaceObj("ModDef", {
-	"title", "All Sponsor Buildings",
+	"title", "All Sponsor Buildings v0.2",
+--~ 	"title", "All Sponsor Buildings",
 	"version", 3,
 	"version_major", 0,
 	"version_minor", 3,