+++ Mods ChoGGi/Fix Canceled Rocket On Pad/Code/Script.lua
@@ -1,0 +1,12 @@
+-- See LICENSE for terms
+
+local IsValid = IsValid
+function OnMsg.LoadGame()
+	local pads = UICity.labels.LandingPad or ""
+	for i = 1, #pads do
+		local pad = pads[i]
+		if pad.rocket_construction and not IsValid(rocket_construction) then
+			pad.rocket_construction = false
+		end
+	end
+end