+++ Mods ChoGGi/Fix Canceled Rocket On Pad/metadata.lua
@@ -1,0 +1,18 @@
+return PlaceObj("ModDef", {
+	"title", "Fix: Canceled Rocket On Pad",
+	"version", 1,
+	"version_major", 0,
+	"version_minor", 1,
+	"saved", 1558872000,
+	"image", "Preview.png",
+	"id", "ChoGGi_FixCanceledRocketOnPad",
+	"steam_id", "1752185403",
+	"pops_any_uuid", "bd3d58a5-dc56-4d3b-9764-2d8fe4de2e76",
+	"author", "ChoGGi",
+	"lua_revision", 244677,
+	"code", {
+		"Code/Script.lua",
+	},
+	"description", [[Start building a rocket on a pad than cancel it, and it's blocked.
+This will check for borked pads when a save is loaded.]],
+})